Reread the requirements and have concluded I was not supposed to be presenting an actual technical article. A subject it appears would be plenty.
So I have removed the blog entries, replaced the button with a
contact me> email link and added a technical subject I am
very interested in.

I also did the ddd_post.html, since it was supposed to be
part of the larger blog context. Finally modified the script and
index yet again and removed unnecessary parts.